MAT-7462 test refactoring and lints
